# perf: optimize Reeds-Shepp path planning with NumPy vectorization